// src/models/schedule.ts
export interface Schedule {
    /**
     * APIVersion defines the versioned schema of this representation of an object.
     * Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values.
     */
    apiVersion?: string;
    /**
     * Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to.
     * Cannot be updated. In CamelCase.
     */
    kind?: string;
    metadata: kubernetes.ObjectMeta;
    spec: ScheduleSpec;
    status: ScheduleStatus;
}
// ...
export interface ScheduleStatus {

    /**
     * Phase a simple, high-level summary of where the schedule is in its lifecycle.
     */
    phase: string;
    startedAt: kubernetes.Time;
    finishedAt: kubernetes.Time;
    /**
     * A human readable message indicating details about why the schedule is in this condition.
     */
    message: string;

    /**
     * Nodes is a mapping between a node ID and the node's status.
     */
    nodes: { [nodeId: string]: NodeStatus};

    /**
     * PersistentVolumeClaims tracks all PVCs that were created as part of the schedule.
     * The contents of this list are drained at the end of the schedule.
     */
    persistentVolumeClaims: kubernetes.Volume[];

    compressedNodes: string;
}

/**
 * ScheduleList is list of Schedule resources
 */
export interface ScheduleList {
    /**
     * APIVersion defines the versioned schema of this representation of an object.
     * Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values.
     */
    apiVersion?: string;
    items: Schedule[];
    /**
     * Kind is a string value representing the REST resource this object represents.
     * Servers may infer this from the endpoint the client submits requests to.
     */
    kind?: string;
    metadata: kubernetes.ListMeta;
}

/**
 * ScheduleSpec is the specification of a Schedule.
 */
export interface ScheduleSpec {
    /**
     * Affinity sets the scheduling constraints for all pods in the schedule. Can be overridden by an affinity specified in the template
     */
    affinity?: kubernetes.Affinity;
    /**
     * Arguments contain the parameters and artifacts sent to the schedule entrypoint.
     * Parameters are referencable globally using the 'schedule' variable prefix. e.g. {{schedule.parameters.myparam}}
     */
    arguments?: Arguments;
    /**
     * Entrypoint is a template reference to the starting point of the schedule
     */
    entrypoint: string;
    /**
     * ImagePullSecrets is a list of references to secrets in the same namespace to use for pulling any images in pods that reference this ServiceAccount.
     * ImagePullSecrets are distinct from Secrets because Secrets can be mounted in the pod, but ImagePullSecrets are only accessed by the kubelet.
     */
    imagePullSecrets?: kubernetes.LocalObjectReference[];

    /**
     * NodeSelector is a selector which will result in all pods of the schedule to be scheduled on the selected node(s).
     * This is able to be overridden by a nodeSelector specified in the template.
     */
    nodeSelector?: { [key: string]: string; };
    /**
     * OnExit is a template reference which is invoked at the end of the schedule, irrespective of the success, failure, or error of the primary schedule.
     */
    onExit?: string;
    /**
     * ServiceAccountName is the name of the ServiceAccount to run all pods of the schedule as.
     */
    serviceAccountName?: string;
    /**
     * Templates is a list of schedule templates used in a schedule
     */
    templates: Template[];
    /**
     * VolumeClaimTemplates is a list of claims that containers are allowed to reference.
     * The Schedule controller will create the claims at the beginning of the schedule and delete the claims upon completion of the schedule
     */
    volumeClaimTemplates?: kubernetes.PersistentVolumeClaim[];
    /**
     * Volumes is a list of volumes that can be mounted by containers in a schedule.
     */
    volumes?: kubernetes.Volume[];
}